Create poj2366两数组10000和数对
Input
You are given both of these lists one by one. Format of each of these lists is as follows: in the first line of the list the quantity of numbers Ni of the i-th list is written. Further there is an i-th list of numbers each number in its line (Ni lines).The following conditions are satisfied: 1 <= Ni <= 50 000, each element of the lists lays in the range from -32768 to 32767. The first list is ascending and the second one is descending.
Output
You should write "YES" to the standard output if it is possible to choose from the two lists of integers such two numbers that their sum would be equal to 10 000. Otherwise you should write "NO".
Sample Input
4
-175
19
19
10424
3
8951
-424
-788
Sample Output
YES
Hint
This problem has huge input data,use scanf() instead of cin to read data to avoid time limit exceed. 

题目大意：
已知两个给定的序列，一个升序排列，一个降序排列，在这两个序列中各找一个数，它们加起来恰好等于10000。

提示：可以二分，也可以hash

 